Forked Search Modules between Apps and Songs
