--- iOSArchitecturesDemo/Flows/Search/Applications/Presenter/SearchPresenter.swift
@@ -1,49 +1,0 @@
-//
-//  SearchPresenter.swift
-//  iOSArchitecturesDemo
-//
-//  Created by Evgeny Kireev on 27/03/2019.
-//  Copyright © 2019 ekireev. All rights reserved.
-//
-
-import UIKit
-
-final class SearchPresenter {
-    
-    weak var viewInput: (UIViewController & SearchViewInput)?
-    
-    private let searchService = ITunesSearchService()
-    
-    private func requestApps(with query: String) {
-        self.searchService.getApps(forQuery: query) { [weak self] result in
-            guard let self = self else { return }
-            self.viewInput?.throbber(show: false)
-            result
-                .withValue { apps in
-                    guard !apps.isEmpty else {
-                        self.viewInput?.showNoResults()
-                        return
-                    }
-                    self.viewInput?.hideNoResults()
-                    self.viewInput?.searchResults = apps
-                }
-                .withError {
-                    self.viewInput?.showError(error: $0)
-                }
-        }
-    }
-}
-
-// MARK: - SearchViewOutput
-extension SearchPresenter: SearchViewOutput {
-    
-    func viewDidSearch(with query: String) {
-        self.viewInput?.throbber(show: true)
-        self.requestApps(with: query)
-    }
-    
-    func viewDidSelectApp(_ app: ITunesApp) {
-        let appDetaillViewController = AppDetailViewController(app: app)
-        self.viewInput?.navigationController?.pushViewController(appDetaillViewController, animated: true)
-    }
-}
