+++ iOSArchitecturesDemo/Flows/Search/View/AppDetail/AppDetailViewController.swift
