+++ iOSArchitecturesDemo/Flows/Search/View/AppDetail/Submodules/AppDetailHeaderViewController.swift
