+++ iOSArchitecturesDemo/Flows/Search/View/AppDetail/Submodules/AppReleaseNotesViewController.swift
