+++ iOSArchitecturesDemo/Flows/Search/View/SearchViewController.swift
@@ -1,0 +1,228 @@
+//
+//  ViewController.swift
+//  iOSArchitecturesDemo
+//
+//  Created by ekireev on 14.02.2018.
+//  Copyright © 2018 ekireev. All rights reserved.
+//
+
+import UIKit
+
+final class SearchViewController: UIViewController {
+  
+  private let presenter: SearchViewOutput
+  
+  // MARK: - Views
+  
+  private let searchBar = UISearchBar()
+  private let tableView = UITableView()
+  private let emptyResultView = UIView()
+  private let emptyResultLabel = UILabel()
+  
+  // MARK: - Properties
+  
+  var searchResults = [Any]() {
+    didSet {
+      self.tableView.isHidden = false
+      self.tableView.reloadData()
+      self.searchBar.resignFirstResponder()
+    }
+  }
+  
+  // MARK: - Private Properties
+  
+  private struct Constants {
+    static let reuseIdentifier = "reuseId"
+  }
+  
+  // MARK: - Init
+  
+  init(presenter: SearchViewOutput) {
+    self.presenter = presenter
+    super.init(nibName: nil, bundle: nil)
+  }
+  
+  required init?(coder aDecoder: NSCoder) {
+    fatalError("init(coder:) has not been implemented")
+  }
+  
+  // MARK: - Lifecycle
+  
+  override func viewDidLoad() {
+    super.viewDidLoad()
+    self.configureUI()
+  }
+  
+  override func viewWillDisappear(_ animated: Bool) {
+    super.viewWillDisappear(animated)
+    self.throbber(show: false)
+  }
+  
+  // MARK: - UI
+  
+  private func configureUI() {
+    self.view.backgroundColor = .white
+    self.navigationController?.navigationBar.prefersLargeTitles = true
+    self.addSearchBar()
+    self.addTableView()
+    self.addEmptyResultView()
+    self.setupConstraints()
+  }
+  
+  private func addSearchBar() {
+    self.searchBar.translatesAutoresizingMaskIntoConstraints = false
+    self.searchBar.delegate = self
+    self.searchBar.searchBarStyle = .minimal
+    self.view.addSubview(self.searchBar)
+  }
+  
+  private func addTableView() {
+    if self.tabBarController?.selectedIndex == 0 {
+      self.tableView.register(AppCell.self, forCellReuseIdentifier: Constants.reuseIdentifier)
+    } else {
+      self.tableView.register(SongCell.self, forCellReuseIdentifier: Constants.reuseIdentifier)
+    }
+    self.tableView.rowHeight = 72.0
+    self.tableView.separatorInset = UIEdgeInsets(top: 0.0, left: 12.0, bottom: 0.0, right: 0.0)
+    self.tableView.translatesAutoresizingMaskIntoConstraints = false
+    self.tableView.delegate = self
+    self.tableView.dataSource = self
+    self.tableView.isHidden = true
+    self.tableView.tableFooterView = UIView()
+    self.view.addSubview(self.tableView)
+  }
+  
+  private func addEmptyResultView() {
+    self.emptyResultView.translatesAutoresizingMaskIntoConstraints = false
+    self.emptyResultView.backgroundColor = .white
+    self.emptyResultView.isHidden = true
+    
+    self.emptyResultLabel.translatesAutoresizingMaskIntoConstraints = false
+    self.emptyResultLabel.text = "Nothing was found"
+    self.emptyResultLabel.textColor = UIColor.darkGray
+    self.emptyResultLabel.textAlignment = .center
+    self.emptyResultLabel.font = UIFont.systemFont(ofSize: 12.0)
+    
+    self.view.addSubview(self.emptyResultView)
+    self.emptyResultView.addSubview(self.emptyResultLabel)
+  }
+  
+  private func setupConstraints() {
+    let safeArea = view.safeAreaLayoutGuide
+    
+    NSLayoutConstraint.activate([
+      self.searchBar.topAnchor.constraint(equalTo: self.view.topAnchor, constant: 8.0),
+      self.searchBar.leadingAnchor.constraint(equalTo: safeArea.leadingAnchor),
+      self.searchBar.trailingAnchor.constraint(equalTo: safeArea.trailingAnchor),
+      
+      self.tableView.topAnchor.constraint(equalTo: self.searchBar.bottomAnchor),
+      self.tableView.leadingAnchor.constraint(equalTo: safeArea.leadingAnchor),
+      self.tableView.trailingAnchor.constraint(equalTo: safeArea.trailingAnchor),
+      self.tableView.bottomAnchor.constraint(equalTo: safeArea.bottomAnchor),
+      
+      self.emptyResultView.topAnchor.constraint(equalTo: self.searchBar.bottomAnchor),
+      self.emptyResultView.leadingAnchor.constraint(equalTo: safeArea.leadingAnchor),
+      self.emptyResultView.trailingAnchor.constraint(equalTo: safeArea.trailingAnchor),
+      self.emptyResultView.bottomAnchor.constraint(equalTo: safeArea.bottomAnchor),
+      self.emptyResultLabel.topAnchor.constraint(equalTo: self.emptyResultView.topAnchor, constant: 12.0),
+      self.emptyResultLabel.leadingAnchor.constraint(equalTo: self.emptyResultView.leadingAnchor),
+      self.emptyResultLabel.trailingAnchor.constraint(equalTo: self.emptyResultView.trailingAnchor)
+      ])
+  }
+}
+
+// MARK: - Input
+
+extension SearchViewController: SearchViewInput {
+  
+  func showError(error: Error) {
+    let alert = UIAlertController(title: "Error", message: "\(error.localizedDescription)", preferredStyle: .alert)
+    let actionOk = UIAlertAction(title: "OK", style: .cancel, handler: nil)
+    alert.addAction(actionOk)
+    self.present(alert, animated: true, completion: nil)
+  }
+  
+  func showNoResults() {
+    self.emptyResultView.isHidden = false
+    self.searchResults = []
+    self.tableView.reloadData()
+  }
+  
+  func hideNoResults() {
+    self.emptyResultView.isHidden = true
+  }
+  
+  func throbber(show: Bool) {
+    UIApplication.shared.isNetworkActivityIndicatorVisible = show
+  }
+}
+
+//MARK: - UITableViewDataSource
+extension SearchViewController: UITableViewDataSource {
+  
+  func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+    return searchResults.count
+  }
+  
+  func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+    let dequeuedCell = tableView.dequeueReusableCell(withIdentifier: Constants.reuseIdentifier, for: indexPath)
+    if self.tabBarController?.selectedIndex == 0 {
+      guard let cell = dequeuedCell as? AppCell, let app = searchResults[indexPath.row] as? ITunesApp else {
+        return dequeuedCell
+      }
+      self.configure(cell: cell, with: app)
+      return cell
+    } else {
+      guard let cell = dequeuedCell as? SongCell, let song = searchResults[indexPath.row] as? ITunesSong else {
+        return dequeuedCell
+      }
+      self.configure(cell: cell, with: song)
+      return cell
+    }
+  }
+  
+  private func configure(cell: AppCell, with app: ITunesApp) {
+    cell.onOpenButtonTap = {
+      guard let urlString = app.appUrl, let url = URL(string: urlString) else { return }
+      UIApplication.shared.open(url, options: [:], completionHandler: nil)
+    }
+    cell.titleLabel.text = app.appName
+    cell.subtitleLabel.text = app.company
+    cell.ratingLabel.text = app.averageRating >>- { "\($0)" }
+  }
+  
+  private func configure(cell: SongCell, with song: ITunesSong) {
+    cell.onSongDetailsButtonTap = {
+      guard let urlString = song.artistViewUrl, let url = URL(string: urlString) else { return }
+      UIApplication.shared.open(url, options: [:], completionHandler: nil)
+    }
+    cell.titleLabel.text = song.trackName
+    cell.artistLabel.text = song.artistName
+  }
+}
+
+//MARK: - UITableViewDelegate
+extension SearchViewController: UITableViewDelegate {
+  
+  func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+    tableView.deselectRow(at: indexPath, animated: true)
+    let item = searchResults[indexPath.row]
+    self.presenter.viewDidSelectItem(item)
+  }
+}
+
+//MARK: - UISearchBarDelegate
+extension SearchViewController: UISearchBarDelegate {
+  
+  func searchBarSearchButtonClicked(_ searchBar: UISearchBar) {
+    guard let query = searchBar.text else {
+      searchBar.resignFirstResponder()
+      return
+    }
+    if query.count == 0 {
+      searchBar.resignFirstResponder()
+      return
+    }
+    self.presenter.viewDidSearch(with: query)
+  }
+}
