+++ iOSArchitecturesDemo/Flows/Search/View/SongCell.swift
@@ -1,0 +1,97 @@
+//
+//  SongCell.swift
+//  iOSArchitecturesDemo
+//
+//  Created by Igor Chernyshov on 02/04/2019.
+//  Copyright © 2019 ekireev. All rights reserved.
+//
+
+import UIKit
+
+final class SongCell: UITableViewCell {
+  
+  private(set) lazy var titleLabel: UILabel = {
+    let label = UILabel()
+    label.translatesAutoresizingMaskIntoConstraints = false
+    label.textColor = .black
+    label.font = UIFont.systemFont(ofSize: 16.0)
+    return label
+  }()
+  
+  private(set) lazy var artistLabel: UILabel = {
+    let label = UILabel()
+    label.translatesAutoresizingMaskIntoConstraints = false
+    label.textColor = .gray
+    label.font = UIFont.systemFont(ofSize: 13.0)
+    return label
+  }()
+  
+  private(set) lazy var songDetailsButton: UIButton = {
+    let button = UIButton(type: .system)
+    button.translatesAutoresizingMaskIntoConstraints = false
+    button.setTitle("Открыть", for: .normal)
+    button.backgroundColor = UIColor(white: 0.9, alpha: 1.0)
+    button.layer.cornerRadius = 16.0
+    button.addTarget(self, action: #selector(songDetailsButtonWasPressed), for: .touchUpInside)
+    return button
+  }()
+  
+  var onSongDetailsButtonTap: (() -> Void)?
+  
+  // MARK: - Init
+  
+  override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
+    super.init(style: style, reuseIdentifier: reuseIdentifier)
+    self.configureUI()
+  }
+  
+  required init?(coder aDecoder: NSCoder) {
+    super.init(coder: aDecoder)
+    self.configureUI()
+  }
+  
+  // MARK: - UI
+  
+  override func prepareForReuse() {
+    [self.titleLabel, self.artistLabel].forEach { $0.text = nil }
+  }
+  
+  private func configureUI() {
+    self.addTitleLabel()
+    self.addArtistLabel()
+    self.addSongDetailsButton()
+  }
+  
+  private func addTitleLabel() {
+    self.contentView.addSubview(self.titleLabel)
+    NSLayoutConstraint.activate([
+      self.titleLabel.topAnchor.constraint(equalTo: self.contentView.topAnchor, constant: 8.0),
+      self.titleLabel.leftAnchor.constraint(equalTo: self.contentView.leftAnchor, constant: 12.0),
+      self.titleLabel.rightAnchor.constraint(equalTo: self.contentView.rightAnchor, constant: -40.0)
+      ])
+  }
+  
+  private func addArtistLabel() {
+    self.contentView.addSubview(self.artistLabel)
+    NSLayoutConstraint.activate([
+      self.artistLabel.topAnchor.constraint(equalTo: self.titleLabel.bottomAnchor, constant: 4.0),
+      self.artistLabel.leftAnchor.constraint(equalTo: self.contentView.leftAnchor, constant: 12.0),
+      self.artistLabel.rightAnchor.constraint(equalTo: self.contentView.rightAnchor, constant: -40.0)
+      ])
+  }
+  
+  private func addSongDetailsButton() {
+    self.contentView.addSubview(self.songDetailsButton)
+    NSLayoutConstraint.activate([
+      self.songDetailsButton.widthAnchor.constraint(equalToConstant: 80.0),
+      self.songDetailsButton.heightAnchor.constraint(equalToConstant: 32.0),
+      self.songDetailsButton.centerYAnchor.constraint(equalTo: self.contentView.centerYAnchor),
+      self.songDetailsButton.rightAnchor.constraint(equalTo: self.contentView.rightAnchor, constant: -12.0)
+      ])
+  }
+  
+  @objc private func songDetailsButtonWasPressed() {
+    onSongDetailsButtonTap?()
+  }
+}
+
